Import of JPRIT/ObjStore-1.42 from CPAN.
gitpan-cpan-distribution: ObjStore
gitpan-cpan-version:      1.42
gitpan-cpan-path:         JPRIT/ObjStore-1.42.tar.gz
gitpan-cpan-author:       JPRIT
gitpan-cpan-maturity:     released

+++ t/zpeek.t
@@ -1,5 +1,5 @@
-# You might need to use GNU diffutils to see differences in -*-perl-*-
-# output due to nulls.
+# You might need to use GNU diff to see differences in -*-perl-*-
+# output due to embedded nulls.
 
 use Test;
 BEGIN { plan tests => 3 }
@@ -50,7 +50,7 @@ $fh->close;
 sub check {
     my ($new,$old) = @_;
     if (-e $old) {
-	if (system("diff $old $new")==0) {
+	if (system("diff -a -c $old $new")==0) {
 	    unlink $new;
 	    ok(1);
 	} else {
